+++ src/dev.js
@@ -1,3 +1,4 @@
+const path = require('path')
 const webpack = require('webpack')
 const merge = require('webpack-merge')
 const base = require('./base')
@@ -9,12 +10,15 @@ module.exports = merge(base, {
     new webpack.HotModuleReplacementPlugin(),
   ],
   devServer: {
-    contentBase: 'dist',
+    contentBase: [path.join(process.cwd(), 'public'), path.join(process.cwd(), 'node_modules')],
     host: '0.0.0.0',
     port: 9001,
     inline: true,
     hot: true,
     disableHostCheck: true,
     historyApiFallback: true,
+    headers: {
+      'Access-Control-Allow-Origin': '*',
+    },
   },
 })