+++ src/output.js
@@ -3,7 +3,7 @@ const path = require('path')
 
 module.exports = {
   path: path.join(process.cwd(), '/dist'),
-  filename: 'js/[name].[conenthash:8].js',
+  filename: '[name].[hash].js',
   publicPath: '/',
-  chunkFilename: 'js/[name].[conenthash:8].js',
+  chunkFilename: '[name].[contenthash].js',
 }