+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 
 ## 0.0.2
 - loader 添加 css 和 less 支持
+
+## 0.0.3
+- 添加 html-webpack-plugin 插件
+- 添加 plugins，entry、output、rules 优化